backend/requirements.txt
fastapi==0.136.0
uvicorn==0.44.0
sqlalchemy==2.0.49
pymysql==1.1.2
python-dotenv==1.2.2
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
pydantic==2.13.2
email-validator==2.3.0
cryptography==46.0.7
numpy==2.4.4
scikit-learn==1.8.0
scipy==1.17.1
